+++ ChangeLog.md
@@ -20,6 +20,7 @@ and this project adheres to
 * NeighborLists and query arguments are now accepted on equal footing by compute methods that involve neighbor finding.
 * 2D PMFTs accept quaternions as well as angles for their orientations.
 * Extensive new documentation including tutorial for new users and reference sections on crucial topics.
+* Added NeighborQuery support to ClusterProperties, GaussianDensity, Voronoi.
 
 ### Changed
 * All compute objects that perform neighbor computations now use NeighborQuery internally.
@@ -44,6 +45,7 @@ and this project adheres to
 * Documentation uses automodule instead of autoclass.
 * The Voronoi class was rewritten to use voro++ for vastly improved performance and correctness in edge cases.
 * MatchEnv has been split into separate classes for the different types of computations it is capable of performing, and these classes all use v2.0-style APIs.
+* Code in the freud.common has been moved to freud.util.
 
 ### Fixed
 * Steinhardt uses the ThreadStorage class and properly resets memory where needed.
@@ -56,6 +58,8 @@ and this project adheres to
 * The freud.util module.
 * Python 2 is no longer supported. Python 3.5+ is required.
 * Cubatic no longer returns the per-particle tensor or the constant r4 tensor.
+* Most features of freud.common are removed from the public API.
+* LinkCell no longer exposes the internals of the cell list data structure.
 
 ## v1.2.2 - 2019-08-15